+++ src/rtp/rtp_packet/rtp_codec.hxx
@@ -40,7 +40,7 @@ class RtpCodec {
   bool marker_ = false;
   uint32_t payload_type_ = 0;
   uint16_t sequence_number_ = 0;
-  uint64_t timestamp_ = 0;
+  uint32_t timestamp_ = 0;
   uint32_t ssrc_ = 0;
   std::vector<uint32_t> csrcs_;
   uint16_t profile_ = 0;
