+++ src/rtp/rtp_packetizer/rtp_packetizer_h264.h
@@ -45,7 +45,7 @@ class RtpPacketizerH264 : public RtpPacketizer {
   bool marker_;
   uint32_t payload_type_;
   uint16_t sequence_number_;
-  uint64_t timestamp_;
+  uint32_t timestamp_;
   uint32_t ssrc_;
   std::vector<uint32_t> csrcs_;
   uint16_t profile_;
